# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResQ Kerala",
    "description": "AI-assisted flood response MVP with human verification and local Ollama agents.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/relief/request": {
      "post": {
        "tags": [
          "Response operations"
        ],
        "summary": "Request Relief",
        "operationId": "request_relief_api_relief_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReliefCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/my": {
      "get": {
        "tags": [
          "Response operations"
        ],
        "summary": "My Relief",
        "operationId": "my_relief_api_relief_my_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/requests": {
      "get": {
        "tags": [
          "Response operations"
        ],
        "summary": "Relief Requests",
        "operationId": "relief_requests_api_relief_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/inventory": {
      "get": {
        "tags": [
          "Response operations"
        ],
        "summary": "Inventory",
        "operationId": "inventory_api_relief_inventory_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Response operations"
        ],
        "summary": "Create Inventory",
        "operationId": "create_inventory_api_relief_inventory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/inventory/{inventory_id}": {
      "put": {
        "tags": [
          "Response operations"
        ],
        "summary": "Update Inventory",
        "operationId": "update_inventory_api_relief_inventory__inventory_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inventory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Inventory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relief/catalog": {
      "get": {
        "tags": [
          "Response operations"
        ],
        "summary": "Catalog",
        "operationId": "catalog_api_relief_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/requests/{request_id}/assign": {
      "post": {
        "tags": [
          "Response operations"
        ],
        "summary": "Assign Request",
        "operationId": "assign_request_api_relief_requests__request_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReliefAssign"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relief/requests/{request_id}/complete": {
      "post": {
        "tags": [
          "Response operations"
        ],
        "summary": "Complete Support",
        "operationId": "complete_support_api_relief_requests__request_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReliefComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relief/distribution": {
      "post": {
        "tags": [
          "Response operations"
        ],
        "summary": "Distribute",
        "operationId": "distribute_api_relief_distribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/relief/distributions": {
      "get": {
        "tags": [
          "Response operations"
        ],
        "summary": "Distributions",
        "operationId": "distributions_api_relief_distributions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DistributionCreate": {
        "properties": {
          "request_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Request Id"
          },
          "inventory_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Inventory Id"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Quantity"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "inventory_id",
          "quantity"
        ],
        "title": "DistributionCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InventoryCreate": {
        "properties": {
          "item": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Item"
          },
          "quantity": {
            "type": "integer",
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Unit",
            "default": "units"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Location"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item",
          "quantity",
          "location"
        ],
        "title": "InventoryCreate"
      },
      "InventoryUpdate": {
        "properties": {
          "item": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Item"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "InventoryUpdate"
      },
...
      "ItemRequest": {
        "properties": {
          "item": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Item"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "inventory_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item",
          "quantity"
        ],
        "title": "ItemRequest"
      },
...
      "ReliefAssign": {
        "properties": {
          "team_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Team Id"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "team_id",
          "note"
        ],
        "title": "ReliefAssign"
      },
      "ReliefComplete": {
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "note"
        ],
        "title": "ReliefComplete"
      },
      "ReliefCreate": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemRequest"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Items"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Location"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          },
          "kind": {
            "type": "string",
            "enum": [
              "supplies",
              "rescue_support"
            ],
            "title": "Kind",
            "default": "supplies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "location"
        ],
        "title": "ReliefCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}